+++ src/main/kotlin/no/nav/familie/ba/sak/behandling/grunnlag/personopplysninger/Person.kt
@@ -109,7 +109,7 @@ data class Person(
     fun hentSeksårsdag(): LocalDate = fødselsdato.plusYears(6)
 }
 
-fun List<Person>.tilBrevTekst(): String = Utils.slåSammen(this.map { it.fødselsdato.tilKortString() })
+fun List<Person>.tilBrevTekst(): String = Utils.slåSammen(this.sortedBy { it.fødselsdato }.map { it.fødselsdato.tilKortString() })
 
 enum class Kjønn {
     MANN,
